requirements.txt
discord.py==2.0.1
python-dotenv==1.0.0
# pipreqs --encoding utf-8 "./"